Revert "Proper initialization of the bridge pot account balances in genesis (#698)"
This reverts commit 311a93046997cfc7738520d88bec44bdd5375fdb.
