+++ crates/bridge-pot-currency-swap/src/existence_optional.rs
@@ -6,14 +6,13 @@ use frame_support::{
     traits::{Currency, ExistenceRequirement, Get, Imbalance, WithdrawReasons},
 };
 
-use super::{Config, CurrencySwap, Pallet};
+use super::{Config, CurrencySwap};
 
 /// A marker type for the implementation that does not require pot accounts existence.
 pub enum Marker {}
 
-impl<T: Config<I>, I: 'static>
-    primitives_currency_swap::CurrencySwap<T::AccountIdFrom, T::AccountIdTo>
-    for CurrencySwap<Pallet<T, I>, Marker>
+impl<T: Config> primitives_currency_swap::CurrencySwap<T::AccountIdFrom, T::AccountIdTo>
+    for CurrencySwap<T, Marker>
 {
     type From = T::CurrencyFrom;
     type To = T::CurrencyTo;
