+++ crates/bridge-pot-currency-swap/src/existence_required.rs
@@ -5,7 +5,7 @@ use frame_support::{
     traits::{Currency, ExistenceRequirement, Get, Imbalance, WithdrawReasons},
 };
 
-use super::{Config, CurrencySwap, Pallet};
+use super::{Config, CurrencySwap};
 
 /// A marker type for the implementation that requires pot accounts existence.
 pub enum Marker {}
@@ -30,9 +30,8 @@ impl From<Error> for DispatchError {
     }
 }
 
-impl<T: Config<I>, I: 'static>
-    primitives_currency_swap::CurrencySwap<T::AccountIdFrom, T::AccountIdTo>
-    for CurrencySwap<Pallet<T, I>, Marker>
+impl<T: Config> primitives_currency_swap::CurrencySwap<T::AccountIdFrom, T::AccountIdTo>
+    for CurrencySwap<T, Marker>
 {
     type From = T::CurrencyFrom;
     type To = T::CurrencyTo;
