+++ crates/humanode-runtime/Cargo.toml
@@ -10,6 +10,7 @@ substrate-wasm-builder = { git = "https://github.com/humanode-network/substrate"
 [dependencies]
 author-ext-api = { version = "0.1", path = "../author-ext-api", default-features = false }
 bioauth-flow-api = { version = "0.1", path = "../bioauth-flow-api", default-features = false }
+bridge-pot-currency-swap = { version = "0.1", path = "../bridge-pot-currency-swap", default-features = false }
 eip191-crypto = { version = "0.1", path = "../eip191-crypto", default-features = false }
 eip191-token-claim = { version = "0.1", path = "../eip191-token-claim", default-features = false }
 eip712-account-claim = { version = "0.1", path = "../eip712-account-claim", default-features = false }
@@ -18,7 +19,6 @@ eip712-token-claim = { version = "0.1", path = "../eip712-token-claim", default-
 keystore-bioauth-account-id = { version = "0.1", path = "../keystore-bioauth-account-id", default-features = false }
 pallet-bioauth = { version = "0.1", path = "../pallet-bioauth", default-features = false }
 pallet-bootnodes = { version = "0.1", path = "../pallet-bootnodes", default-features = false }
-pallet-bridge-pot-currency-swap = { version = "0.1", path = "../pallet-bridge-pot-currency-swap", default-features = false }
 pallet-chain-properties = { version = "0.1", path = "../pallet-chain-properties", default-features = false }
 pallet-chain-start-moment = { version = "0.1", path = "../pallet-chain-start-moment", default-features = false }
 pallet-currency-swap = { version = "0.1", path = "../pallet-currency-swap", default-features = false }
@@ -138,6 +138,7 @@ runtime-benchmarks = [
 std = [
   "author-ext-api/std",
   "bioauth-flow-api/std",
+  "bridge-pot-currency-swap/std",
   "chrono/std",
   "codec/std",
   "eip191-crypto/std",
@@ -161,7 +162,6 @@ std = [
   "pallet-base-fee/std",
   "pallet-bioauth/std",
   "pallet-bootnodes/std",
-  "pallet-bridge-pot-currency-swap/std",
   "pallet-chain-properties/std",
   "pallet-chain-start-moment/std",
   "pallet-currency-swap/std",
@@ -218,6 +218,7 @@ std = [
   "vesting-scheduling-timestamp/std",
 ]
 try-runtime = [
+  "bridge-pot-currency-swap/try-runtime",
   "fp-self-contained/try-runtime",
   "frame-executive/try-runtime",
   "frame-support/try-runtime",
@@ -229,7 +230,6 @@ try-runtime = [
   "pallet-base-fee/try-runtime",
   "pallet-bioauth/try-runtime",
   "pallet-bootnodes/try-runtime",
-  "pallet-bridge-pot-currency-swap/try-runtime",
   "pallet-chain-properties/try-runtime",
   "pallet-chain-start-moment/try-runtime",
   "pallet-currency-swap/try-runtime",
