+++ crates/humanode-runtime/src/currency_swap.rs
@@ -1,22 +1,44 @@
-use pallet_bridge_pot_currency_swap::ExistenceRequired;
+use bridge_pot_currency_swap::ExistenceRequired;
+use sp_runtime::traits::Identity;
 
 use crate::{
-    AccountId, EvmAccountId, EvmToNativeSwapBridge, EvmToNativeSwapBridgePot, FeesPot,
-    NativeToEvmSwapBridge, TreasuryPot,
+    parameter_types, AccountId, Balances, EvmAccountId, EvmBalances, EvmToNativeSwapBridgePot,
+    FeesPot, NativeToEvmSwapBridgePot, TreasuryPot,
 };
 
+parameter_types! {
+    pub NativeToEvmSwapBridgePotAccountId: AccountId = NativeToEvmSwapBridgePot::account_id();
+    pub EvmToNativeSwapBridgePotAccountId: EvmAccountId = EvmToNativeSwapBridgePot::account_id();
+}
+
 pub type NativeToEvmOneToOne =
-    pallet_bridge_pot_currency_swap::CurrencySwap<NativeToEvmSwapBridge, ExistenceRequired>;
+    bridge_pot_currency_swap::CurrencySwap<NativeToEvmOneToOneConfig, ExistenceRequired>;
+
+pub struct NativeToEvmOneToOneConfig;
+
+impl bridge_pot_currency_swap::Config for NativeToEvmOneToOneConfig {
+    type AccountIdFrom = AccountId;
+    type AccountIdTo = EvmAccountId;
+    type CurrencyFrom = Balances;
+    type CurrencyTo = EvmBalances;
+    type BalanceConverter = Identity;
+    type PotFrom = NativeToEvmSwapBridgePotAccountId;
+    type PotTo = EvmToNativeSwapBridgePotAccountId;
+}
 
 pub type EvmToNativeOneToOne =
-    pallet_bridge_pot_currency_swap::CurrencySwap<EvmToNativeSwapBridge, ExistenceRequired>;
+    bridge_pot_currency_swap::CurrencySwap<EvmToNativeOneToOneConfig, ExistenceRequired>;
 
-pub struct GenesisVerifier;
+pub struct EvmToNativeOneToOneConfig;
 
-impl pallet_bridge_pot_currency_swap::GenesisVerifier for GenesisVerifier {
-    fn verify() -> bool {
-        true
-    }
+impl bridge_pot_currency_swap::Config for EvmToNativeOneToOneConfig {
+    type AccountIdFrom = EvmAccountId;
+    type AccountIdTo = AccountId;
+    type CurrencyFrom = EvmBalances;
+    type CurrencyTo = Balances;
+    type BalanceConverter = Identity;
+    type PotFrom = EvmToNativeSwapBridgePotAccountId;
+    type PotTo = NativeToEvmSwapBridgePotAccountId;
 }
 
 pub struct EvmToNativeProxyConfig;
