+++ crates/humanode-runtime/src/tests/claims_and_vesting.rs
@@ -353,9 +353,7 @@ fn prepare_genesis_json(token_claims: &str, token_claim_pot_balance: u128) -> St
         }},
         "evmToNativeSwapBridgePot": {{
             "initialState": "Initialized"
-        }},
-        "nativeToEvmSwapBridge": null,
-        "evmToNativeSwapBridge": null
+        }}
     }}"#
     )
 }
