+++ crates/humanode-runtime/src/tests/genesis_config.rs
@@ -117,9 +117,7 @@ fn works() {
         },
         "evmToNativeSwapBridgePot": {
             "initialState": "Initialized"
-        },
-        "nativeToEvmSwapBridge": null,
-        "evmToNativeSwapBridge": null
+        }
     }"#;
     let config: GenesisConfig = serde_json::from_str(json_input).unwrap();
     assert_ok!(config.build_storage());
@@ -138,8 +136,7 @@ fn unknown_field() {
         `system`, `bootnodes`, `bioauth`, `babe`, `balances`, `treasuryPot`, \
         `feesPot`, `tokenClaimsPot`, `transactionPayment`, `session`, `chainProperties`, \
         `ethereumChainId`, `sudo`, `grandpa`, `ethereum`, `evm`, `dynamicFee`, `baseFee`, \
-        `imOnline`, `evmAccountsMapping`, `tokenClaims`, `nativeToEvmSwapBridgePot`, \
-        `evmToNativeSwapBridgePot`, `nativeToEvmSwapBridge`, `evmToNativeSwapBridge` at line 1 column 6"
+        `imOnline`, `evmAccountsMapping`, `tokenClaims`, `nativeToEvmSwapBridgePot`, `evmToNativeSwapBridgePot` at line 1 column 6"
     );
 }
